feat: implement rechunk (#21)
* feat: implement rechunk

You still have to manually load the image.

Signed-off-by: m2 <69128853+m2Giles@users.noreply.github.com>

* subfolder per image in build-dir

Signed-off-by: m2 <69128853+m2Giles@users.noreply.github.com>

* add pull-retry

These pull-retry's will not be needed if we build inside the devcontainter.

This is a workaround for the 24.04 runner not having a new enough podman that includes retry. Skopeo could be used if we fix the AppArmor or whatever is blocking the namespace.

Signed-off-by: m2 <69128853+m2Giles@users.noreply.github.com>

* nest the workspace

Signed-off-by: m2 <69128853+m2Giles@users.noreply.github.com>

* don't use `

Signed-off-by: m2 <69128853+m2Giles@users.noreply.github.com>

---------

Signed-off-by: m2 <69128853+m2Giles@users.noreply.github.com>