Remove CPP warnings.
This removes the remaining cpp warnings by removing the following.

#include scripts do not have a shebang line.

The shellcheck extension is removed in favor of only using the bash-ide extension.

That extension is configured to ignore SC2148 and SC1113 which handles shebang lines. bash-ide extension already adds the --shell=bash flag to shellcheck (and shellcheck is in the dev container) meaning we do not need the duplicative ShellCheck extension.

Downsides: This workaround is designed specifically with vscode's devcontainer solution in mind. We could add a .dir-local and .vscode, but this is definitely an odd workaround.

We could not do shebang removal and recognize that the cpp warnings should only ever be that first line.

Signed-off-by: m2 <69128853+m2Giles@users.noreply.github.com>