refactor: merge NAS into base, reorganize scripts (#38)
* refactor: merge NAS into base, reorganize scripts

Signed-off-by: m2 <69128853+m2Giles@users.noreply.github.com>

* use includes instead of COPY

Signed-off-by: m2 <69128853+m2Giles@users.noreply.github.com>

* fix: use server in description instead of cayo

Signed-off-by: m2 <69128853+m2Giles@users.noreply.github.com>

* don't include "server" in HCI descriptions

Signed-off-by: m2 <69128853+m2Giles@users.noreply.github.com>

* wrap comments with C style comments.

This definitely is not ergonomic, but it removes the vast majority of the Warnings.

Signed-off-by: m2 <69128853+m2Giles@users.noreply.github.com>

* Consistent style

Signed-off-by: m2 <69128853+m2Giles@users.noreply.github.com>

* Remove CPP warnings.

This removes the remaining cpp warnings by removing the following.

#include scripts do not have a shebang line.

The shellcheck extension is removed in favor of only using the bash-ide extension.

That extension is configured to ignore SC2148 and SC1113 which handles shebang lines. bash-ide extension already adds the --shell=bash flag to shellcheck (and shellcheck is in the dev container) meaning we do not need the duplicative ShellCheck extension.

Downsides: This workaround is designed specifically with vscode's devcontainer solution in mind. We could add a .dir-local and .vscode, but this is definitely an odd workaround.

We could not do shebang removal and recognize that the cpp warnings should only ever be that first line.

Signed-off-by: m2 <69128853+m2Giles@users.noreply.github.com>

* remove unnecessary echo

Signed-off-by: m2 <69128853+m2Giles@users.noreply.github.com>

* remove exectuable bit from #include scripts

Signed-off-by: m2 <69128853+m2Giles@users.noreply.github.com>

* configure shellcheck instead of using bash-ide extension

Signed-off-by: m2 <69128853+m2Giles@users.noreply.github.com>

* provide comments to the configuration

Signed-off-by: m2 <69128853+m2Giles@users.noreply.github.com>

* formatting and .dir-locals for emacs

Signed-off-by: m2 <69128853+m2Giles@users.noreply.github.com>

* remove executable bit from includes

Signed-off-by: m2 <69128853+m2Giles@users.noreply.github.com>

---------

Signed-off-by: m2 <69128853+m2Giles@users.noreply.github.com>