feat: do not include candidates for shipping via extension
The team has discussed creating an extension model similar to that used
by Flatcar Linux as a means to ship features which are highly desireable
but also subject to specific use cases. Any candidate for shipping via
extension should not be included in the base Cayo image to avoid
creating later problems in transitioning from one approach to the other.

The following lists all packages previously installed in the base image
but now marked as an extension candidate:
- docker-ce
- intel-compute-runtime
- samba
- wifi support

As of this commit, these packages are still installable via there new
numbered build scripts, should one chose to modify the Containerfile.in.
Keeping them here now will aid in extension developement, but they
should not be included in a built image.

NVIDIA and libvirt are already gated behind conditional build logic, so
I've left them as is. However, our initial workflows should NOT build
and publish HCI or NVIDIA variant images.

Additionally, small cleanups were made to os-release modification and
base package install order.