fix: remove zfs-dracut to prevent errors (#77)
zfs-dracut is included with the initramfs rebuild, which includes a
default hostname and sets a default hostid from ZFS's perspective, so
when this boots on a ZFS system, a zfs service fails during the
initramfs boot stage due to the hostname of the booting system not
matching the actual system hostname.

There are some ways this could be addressed, but really, we don't need
zfs-dracut because root on ZFS is not yet working for bootc installs and
we don't have planned support for it at this time.